requirements.txt
Flask
Flask-Cors
Flask-Session
Flask-Limiter
python-dotenv
bcrypt
cryptography
redis
gunicorn
whitenoise